--- a/test/studies/shootout/nbody/bradc/nbody-blc.chpl
+++ b/test/studies/shootout/nbody/bradc/nbody-blc.chpl
@@ -1,15 +1,15 @@
 /* The Computer Language Benchmarks Game
    https://salsa.debian.org/benchmarksgame-team/benchmarksgame/
 
-   contributed by Albert Sidelnik and Brad Chamberlain
-   derived from the Java version by Mark C. Lewis and Chad Whipkey
+   contributed by Brad Chamberlain
+   derived from the Chapel version by Albert Sidelnik and myself
 */
 
 
 config const n = 10000;       // The number of timesteps to simulate
 
-const pi = 3.141592653589793,
-      solarMass = 4 * pi**2,
+param pi = 3.141592653589793,
+      solarMass = 4 * pi * pi,
       daysPerYear = 365.24;
 
 
@@ -24,9 +24,9 @@ record body {
 }
 
 //
-// the array of bodies that we'll be simulating
+// the bodies that we'll be simulating
 //
-var bodies = [/* sun */
+var bodies = (/* sun */
               new body(mass = solarMass),
 
               /* jupiter */
@@ -64,10 +64,9 @@ var bodies = [/* sun */
                                1.62824170038242295e-03 * daysPerYear,
                               -9.51592254519715870e-05 * daysPerYear),
                       mass =   5.15138902046611451e-05 * solarMass)
-              ];
+              );
 
-const numBodies = bodies.size;    // the number of bodies being simulated
-
+param numBodies = 5;
 
 proc main() {
   initSun();                      // initialize the sun's velocity
@@ -92,21 +91,20 @@ proc initSun() {
 // advance the positions and velocities of all the bodies
 //
 proc advance(dt) {
-  for i in 0..<numBodies {
-    for j in i+1..<numBodies {
-      ref b1 = bodies[i],
-          b2 = bodies[j];
+  for param i in 0..<numBodies {
+    for param j in i+1..<numBodies {
+      const dpos = bodies[i].pos - bodies[j].pos,
+            sumOfSq = sumOfSquares(dpos),
+            mag = dt / (sumOfSq*sqrt(sumOfSq));
 
-      const dpos = b1.pos - b2.pos,
-            mag = dt / sqrt(sumOfSquares(dpos))**3;
-
-      b1.vel -= dpos * b2.mass * mag;
-      b2.vel += dpos * b1.mass * mag;
+      bodies[i].vel -= dpos * bodies[j].mass * mag;
+      bodies[j].vel += dpos * bodies[i].mass * mag;
     }
   }
 
-  for b in bodies do
-    b.pos += dt * b.vel;
+  for param i in 0..<numBodies {
+    bodies[i].pos += dt * bodies[i].vel;
+  }
 }
 
 //
@@ -116,14 +114,10 @@ proc energy() {
   var e = 0.0;
 
   for i in 0..<numBodies {
-    const b1 = bodies[i];
-
-    e += 0.5 * b1.mass * sumOfSquares(b1.vel);
-
+    e += 0.5 * bodies[i].mass * sumOfSquares(bodies[i].vel);
     for j in i+1..<numBodies {
-      const b2 = bodies[j];
-
-      e -= (b1.mass * b2.mass) / sqrt(sumOfSquares(b1.pos - b2.pos));
+      e -= (bodies[i].mass * bodies[j].mass)
+           / sqrt(sumOfSquares(bodies[i].pos - bodies[j].pos));
     }
   }