@ronawho 2022-03-24
Here's the performance I see on the shootout like machine:

```sh
chpl test/studies/shootout/nbody/bradc/nbody-blc.chpl --fast
time -p ./nbody-blc --n=50000000 >/dev/null
```

| Config | Time  |
| ------ | ----: |
| before | 4.95s |
| now    | 3.70s |


And checking which components make the biggest difference:

| Config          | Time  |
| --------------- | ----: |
| before          | 4.95s |
| param pi        | 4.95s |
| tuple bodies    | 4.22s |
| param numBodies | 4.06s |
| param loops     | 3.91s |
| sumOfSq trick   | 3.70s |
| now             | 3.70s |

So sumOfSq trick helps, but we're doing less work than other versions, so I suspect it may not be allowed. Getting rid of the refs does not seem to change performance, so I don't see any reason to keep them around.