@bradcray 2022-03-24
>  but we're doing less work than other versions

I don't think it's less work.  Basically, we want to compute `sqrt(n)**3` == `n**(3/2)` == `n*(n**1/2)` == `n*sqrt(n)`

> so I suspect it may not be allowed

I got it from one of the Julia versions, so if he disallowed it, they'd need to rewrite as well.  But given that it only helps a little compared to the other changes and I'm not fond of it stylistically, I'd say let's start with the cleaner version and then only resort to that if we are so close that we think it'd push us into the lead.