@bradcray 2022-03-24
@chapel-lang/perf-team : memleak, valgrind, and asan testing came back clean, as expected.