@ronawho 2022-03-25
FYI my original results were with llvm-11, but I see no perf difference with llvm-13.

> gets rid of the refs (used to be crucial for perf, here don't seem to be)

Discussed offline, but we think these were most important when DR array indexing was slower. We also guessed that it probably makes no difference with the switch here to param loops. 

I was curious about the impact on main, and was surprised to find getting rid of the refs seems to improve performance. I see 4.95s on main, and 4.75s with refs in `advance` removed.