@ronawho 2022-03-25
Note that your "now" timings in the PR are off since they reflect the timings with the sumOfSq trick.